requirements.txt
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.0.1
pandas==2.1.3
googletrans==4.0.0rc1
openpyxl==3.1.2
tk==0.1.0